requirements.txt
nautilus-trader[bybit]
pandas
numpy
matplotlib
jupyter
python-dotenv